Move dilute tests into own file
